use 3d grid to optimize 3d simulation
