+++ src/simulation/simulation_2d.hpp
@@ -3,4 +3,4 @@
 #include "simulation.hpp"
 #include "spatial_grid.hpp"
 
-void simulate_2d(float dt, cgp::numarray<particle_element>& particles, spatial_grid& grid, sph_parameters_structure const& sph_parameters);
+void simulate_2d(float dt, cgp::numarray<particle_element>& particles, spatial_grid_base& grid, sph_parameters_structure const& sph_parameters);
