+++ src/simulation/simulation_3d.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
 #include "simulation.hpp"
+#include "spatial_grid.hpp"
 
-void simulate_3d(float dt, cgp::numarray<particle_element>& particles, sph_parameters_structure const& sph_parameters);
+void simulate_3d(float dt, cgp::numarray<particle_element>& particles, spatial_grid_base& grid, sph_parameters_structure const& sph_parameters);
